Progress on migration to ImgLib2:
  * Renamed all package prefixes to net.imglib2.
  * Updated imagej.data.Metadata.
  * Began updating various other classes.


git-svn-id: http://code.imagej.net/svn/imagej/branches/imglib2@2694 be23795f-a60e-40d4-ba19-055798872e57

+++ core/display/src/main/java/imagej/display/DisplayController.java
@@ -49,9 +49,9 @@ public interface DisplayController {
 
 	AxisLabel[] getDimLabels();
 
-	int[] getDims();
+	long[] getDims();
 
-	int[] getPos();
+	long[] getPos();
 
 	Dataset getDataset();
